requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
openai>=1.0.0
python-dotenv==1.0.1
